+++ src/bin/flora-extension.ts
@@ -42,9 +42,9 @@ program
 program
   .command("publish")
   .description(
-    "Create an extensions.json entry for a released extension. This can be added to the https://github.com/foxglove/extension-registry repository",
+    "Create a marketplace entry for a released Flora extension",
   )
-  .option("--foxe <foxe>", "URL of the published .foxe file")
+  .option("--foxe <foxe>", "HTTPS URL of the published .foxe file")
   .option("--cwd [cwd]", "Directory containing the extension package.json file")
   .option("--version [version]", "Version of the published .foxe file")
   .option("--readme [readme]", "URL of the extension README.md file")
@@ -53,7 +53,7 @@ program
     main(publishCommand(options));
   });
 
-program.on("command:*", ([_cmd]: string) => {
+program.on("command:*", () => {
   program.outputHelp({ error: true });
   process.exit(1);
 });