+++ examples/panel-settings/CHANGELOG.md
@@ -6,4 +6,4 @@
 
 ## 0.0.1
 
-- Updated to use latest version of `create-foxglove-extension` and fixes broken dependency (`react-json-view` -> `@microlink/react-json-view`).
+- Updated to use latest version of `@flora-suite/create-flora-extension` and fixes broken dependency (`react-json-view` -> `@microlink/react-json-view`).